feat: support custom delimiter for cookie value separation
When storing the value in the cookie both the non-hashed token and the
hashed token are concatenated together separated by a delimiter. This
is so that stateless re-use of the token value can be supported.

Previously the delimiter was hardcoded to a '|' character, this can now
be overridden via the delimiter configuration option.