feat: re-allow httpOnly override
Technically there's nothing wrong with disabling the httpOnly option
for the csrf cookie. It's a valid way for the frontend to also extract
the token value to then include it in a header or request body, many
other CSRF solutions allow this.

Due to some controversial discussions around a particular use case
where Twitter had a CSRF vulnerability from a vulnerable subdomain. The
httpOnly option being set to true could have mitigated the vulnerability
in that particular case. Because of this, csrf-csrf enforced a httpOnly
flag of true. However, it should be up to the developer to determine
the risk involved and whether or not that's the implementation they
want to use.

csrf-csrf will default to httpOnly true and will continue to recommend
keeping it as true, but will not restrict the choice to balance their
security as they feel fit.