feat!: change default value of sameSite to 'strict'
BREAKING CHANGE: If you were previously dependent on the
sameSite value being 'lax' and you weren't explicitly setting it
then you will need to set it now.

The reason for this change is to try and make csrf-csrf more
secure by default. The idea being that when you need to make
a change to the configuration, you should be actively thinking
about that change and determining whether it is correct. This
is to help encourage people to consult the documentation
regarding configuration considerations.