feat: set httpOnly default back to true
This is non-breaking as it reverts a previously unreleased
breaking change to false. The change means less breaking
changes with v4, less risk, and the WIP v4 documentation
is intended to try and better guide users around appropriate
configuration.

The idea is that users should be properly determining whether
a particular config change is appropriate, they should make
sure they understand what the configuration does and what
impact changing it may have.