feat!: change validateOnReuse to false by default
BREAKING CHANGE: If you were previously relying on either
the generateToken or req.csrfToken functions throwing an
error when the current CSRF token cookie is invalid, you will
need to manually set validateOnReuse to false.

An error will no longer be thrown by default, a new token will
be generated and returned instead. If you previously had this
manually set to false, this is no longer necessary as it is now
the default value.